clipclap/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
open-clip-torch>=2.24.0
laion-clap>=1.1.4
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
